requirements.txt
Django==4.2
pillow
django-cleanup
django-allauth
django-htmx